+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Improved performance of RoadIntersectionEditor Init function
 - Refactored CheckLoadTexture into EditorUtilities
 - Removed unused editor timer
+- Unified DrawLine in EditorUtilities
 
 
 ## [Version: 1.9_FH] - 2019.04.22